requirements.txt
PySide6
tomlkit
bip-utils
requests
mysql-connector-python
qrcode